+++ src/jit/xm_codegen_x64_ins.c
@@ -1493,7 +1493,7 @@ static void x64_h_suspend(X64CodegenCtx *ctx, XmIns *ins, X64Reg rd) {
     /* Load result_tag -> runtime_tags[bc_slot] */
     {
         int32_t res_vreg_off = -1;
-        int16_t res_bc_slot = -1;
+        int32_t res_bc_slot = -1;
         if (xm_ref_is_vreg(ins->dst)) {
             uint32_t vi = XM_REF_INDEX(ins->dst);
             if (vi < ctx->func->nvreg && vi < XR_JIT_MAX_VREG_TAGS)
